+++ mybackupy/__init__.py
@@ -1,3 +1,11 @@
 __version__ = "1.0.0b0"
 __license__ = "MIT"
 __title__ = "mybackupy"
+
+from mybackupy.cli import Cli
+
+def main():
+    Cli()
+
+if __name__ == "__main__":
+    main()
